Refactored the events logic to mirror CTA stuff (#16)
Merging to master.
